+++ tests/end_to_end/target_snowflake/tap_postgres/test_resync_pg_to_sf_with_split_large_files.py
@@ -2,7 +2,7 @@ from pipelinewise.fastsync import postgres_to_snowflake
 from tests.end_to_end.helpers import assertions
 from tests.end_to_end.target_snowflake.tap_postgres import TapPostgres
 
-TAP_ID = 'postgres_to_sf_split_large_files'
+TAP_ID = 'pg_to_sf_split_large_files'
 TARGET_ID = 'snowflake'